# repository/index: Avoid creating an error when checking if an id is in the index.