@MJDSys 2018-01-08
Regarding the benchmark, I posted a note about the speed up I get running a backup of my system (saving about an hour of time in a 2-3 hour backup).  Is there a more established way to do the benchmark?  If it helps, I found this by doing some CPU profiling of my backup.

Regarding your suggestion, that does sound great.  If I understand correctly, I just have the one outer if statement?  That sounds like it could help some more (lookupInternal still dominates the cpu profile).  I'll give it a test here and see how it goes.