@fd0 2018-01-08
Hey, for Go software it's very easy to write a micro benchmark of just the `Has()` function for the index:
```go
const maxPackSize = 16 * 1024 * 1024

func createRandomIndex() (idx *repository.Index, lookupID restic.ID) {
	idx = repository.NewIndex()

	// create index with 200k pack files
	for i := 0; i < 200000; i++ {
		packID := restic.NewRandomID()
		offset := 0
		for offset < maxPackSize {
			size := 2000 + rand.Intn(4*1024*1024)
			id := restic.NewRandomID()
			idx.Store(restic.PackedBlob{
				PackID: packID,
				Blob: restic.Blob{
					Type:   restic.DataBlob,
					ID:     id,
					Length: uint(size),
					Offset: uint(offset),
				},
			})

			offset += size

			if rand.Float32() < 0.001 && lookupID.IsNull() {
				lookupID = id
			}
		}
	}

	return idx, lookupID
}

func BenchmarkIndexHasUnknown(b *testing.B) {
	idx, _ := createRandomIndex()
	lookupID := restic.NewRandomID()

	b.ResetTimer()

	for i := 0; i < b.N; i++ {
		idx.Has(lookupID, restic.DataBlob)
	}
}

func BenchmarkIndexHasKnown(b *testing.B) {
	idx, lookupID := createRandomIndex()

	b.ResetTimer()

	for i := 0; i < b.N; i++ {
		idx.Has(lookupID, restic.DataBlob)
	}
}
```

You can then run just this benchmark and look at the memory allocation:
```
$ go test -v -run xxxxx -bench IndexHas -benchmem | tee before
goos: linux
goarch: amd64
pkg: github.com/restic/restic/internal/repository
BenchmarkIndexHasUnknown-4   	 1000000	      1464 ns/op	     608 B/op	      10 allocs/op
BenchmarkIndexHasKnown-4     	 5000000	       280 ns/op	     128 B/op	       5 allocs/op
PASS
ok  	github.com/restic/restic/internal/repository	20.428s
```

So, we can see that each call of `Has()` takes about 300ns and allocates 128 Byte for known blobs, and 1500ns and 600 byte for unknown blobs.

With your change cherry picked, this looks already much better:

```
$ go test -v -run xxxxx -bench IndexHas -benchmem | tee after
goos: linux
goarch: amd64
pkg: github.com/restic/restic/internal/repository
BenchmarkIndexHasUnknown-4   	10000000	       143 ns/op	      16 B/op	       2 allocs/op
BenchmarkIndexHasKnown-4     	 5000000	       276 ns/op	     128 B/op	       5 allocs/op
PASS
ok  	github.com/restic/restic/internal/repository	22.466s
```

We can use `benchcmp` to compare it:

```
$ benchcmp before after
benchmark                      old ns/op     new ns/op     delta
BenchmarkIndexHasUnknown-4     1464          143           -90.23%
BenchmarkIndexHasKnown-4       280           276           -1.43%

benchmark                      old allocs     new allocs     delta
BenchmarkIndexHasUnknown-4     10             2              -80.00%
BenchmarkIndexHasKnown-4       5              5              +0.00%

benchmark                      old bytes     new bytes     delta
BenchmarkIndexHasUnknown-4     608           16            -97.37%
BenchmarkIndexHasKnown-4       128           128           +0.00%
```

With my suggestions, we can even improve further:

```
$ go test -v -run xxxxx -bench IndexHas -benchmem | tee after2
goos: linux
goarch: amd64
pkg: github.com/restic/restic/internal/repository
BenchmarkIndexHasUnknown-4   	20000000	        76.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkIndexHasKnown-4     	20000000	        72.9 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/restic/restic/internal/repository	22.371s
```

And benchcmp agrees: :smile:
```
$ benchcmp before after2
benchmark                      old ns/op     new ns/op     delta
BenchmarkIndexHasUnknown-4     1464          76.4          -94.78%
BenchmarkIndexHasKnown-4       280           72.9          -73.96%

benchmark                      old allocs     new allocs     delta
BenchmarkIndexHasUnknown-4     10             0              -100.00%
BenchmarkIndexHasKnown-4       5              0              -100.00%

benchmark                      old bytes     new bytes     delta
BenchmarkIndexHasUnknown-4     608           0             -100.00%
BenchmarkIndexHasKnown-4       128           0             -100.00%
```

So, if it's okay for you, I'd like to take over from here. Thank you very very much for pointing me to this improvement opportunity!