@MJDSys 2018-01-08
Ah, I didn't realize you meant that kind of benchmark.  Sorry about that.

Feel free to take over this pr if you want.  I pushed my changes to the Has function (undoing my other changes).  I also whipped up a quick test for the Has function (though it seems thoroughly tested by all the other tests), so feel free to grab that as well.  I have checked the "Allow Maintainers to make changes" button if you want to just add to this pr.

I think there is still some optimizations available to get my backup to run faster, so I'll look into those as well.  If I find any, I'll try to add a mirco-benchmark for those as well next time.